backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pydantic==2.10.6
firebase-admin==6.6.0
python-dotenv==1.0.1
cors==1.0.1
google-generativeai==0.8.4
